Further failover logic by tracking backup server keys
If a primary fails and a backup takes over, that new
primary needs to continue logging transactions associated
with the failed primary to the other backups of said primary.
However, the new primary may already have been storing
a set of keys, which have a different set of backups.

Added logic to track which backup servers are associated
with which keys. So after failover, on log_put calls,
the new primary knows which servers to backup keys to.

Signed-off-by: Cody D'Ambrosio <cjd218@lehigh.edu>