Add one-sided RDMA logging and fix recovery bug
Add build define for one-sided RDMA logging, enabled
by default. Non-optimal though as only one write
can be performed before waiting for backup to read it.

Fix failover issue. If a backup took over as primary,
it would open an endpoint for the original primary to
connect to. Then if the primary came back online, it came
up as a backup. If the new primary (originally backup) failed,
the original primary would become primary again. However, the
original primary would not be able to connect to the backup
if it came back online. Added tracking of original server
states to handle this.

Signed-off-by: Cody D'Ambrosio <cjd218@lehigh.edu>