Fix failover logic for circular backups
When heartbeat to backup dies, the backup that died
may have original been a primary and will come back up as such.
Open endpoint to accept it.

The server backing us up may have other servers it is backing up to.
But those would be for different keys and we would not take over
for those keys, so remove from backup list.

Signed-off-by: Cody D'Ambrosio <cjd218@lehigh.edu>